'main'

+++ src/App.tsx
@@ -1,6 +1,7 @@
+import React from "react";
 import Home from "./components/pages/Home/Home";
 import "./index.scss";
-import { Route, Routes } from "react-router-dom";
+import { Route, Routes, useNavigate } from "react-router-dom";
 import Cart from "./components/pages/Cart/Cart";
 import { AppContext } from "./context/AppContext";
 import { useDispatch } from "react-redux";
@@ -8,16 +9,31 @@ import { fetchCards } from "./redux/actions/cards";
 import { useEffect, useState } from "react";
 import axios from "axios";
 import Orders from "./components/pages/Orders/Orders";
+import ItemDescription from "./components/pages/Home/ItemDescription/ItemDescription";
+import { string } from "prop-types";
 //import cards from "./redux/redusers/cards";
 //import filterItems from "./redux/redusers/filterItems";
 //import store from "./redux/store";
 
-function App() {
+const App:React.FC=()=> {
 
-  const dispatch = useDispatch();
+  const dispatch:Function = useDispatch();
   const [isLoading, setIsLoading] = useState(false);
-  const [cartItems, setCartItems] = useState([]);
+  type CartItemsType={
+    id:string|undefined;
+    car:string|undefined;
+    stockPrice:string|undefined;
+    tunerPrice:string|undefined;
+    parentId:number|undefined;
+    choosedType:boolean|undefined;
+    stockImage:string|undefined;
+    tunerImage:string|undefined;
+    stockHP:string|undefined;
+    tunerHP:string|undefined;
+  }[];
+  const [cartItems, setCartItems] = useState<CartItemsType>();
 
+  const navigate = useNavigate();
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -31,58 +47,61 @@ function App() {
         setIsLoading(false);
       } catch (error) {
         console.log("Error with load data: ", error);
+        navigate('/')
       }
     };
     fetchData();
   }, []);
 
-  const onAddToCart = async (obj) => {
+  const onAddToCart = async (obj: {id:string|null}) => {
     try {
-      const findItem = cartItems.find(
-        (el) => Number(el.parentId) === Number(obj.id)
-      );
-      if (findItem) {
-        setCartItems((prev) =>
-          prev.filter((el) => Number(el.parentId) !== Number(obj.id))
-        );
-        await axios.delete(
-          `https://631076b736e6a2a04eeef849.mockapi.io/cartItems/${findItem.id}`
-        );
-      } else {
-        setCartItems((prev) => [...prev, obj]);
-        const { data } = await axios.post(
-          "https://631076b736e6a2a04eeef849.mockapi.io/cartItems",
-          obj
-        );
-        setCartItems((prev) =>
-          prev.map((item) => {
-            if (item.parentId === data.parentId) {
-              return {
-                ...item,
-                id: data.id,
-              };
-            }
-            return item;
-          })
-        );
+      if(cartItems){
+        const findItem:{id:string|undefined}|undefined=cartItems.find(
+          (el) => Number(el.parentId) === Number(obj.id)
+        )
+        if (findItem) {
+          setCartItems((prev:object|any) =>
+            prev.filter((el:object|any) => Number(el.parentId) !== Number(obj.id))
+          );
+          await axios.delete(
+            `https://631076b736e6a2a04eeef849.mockapi.io/cartItems/${findItem.id}`
+          );
+        } else {
+          setCartItems((prev:any) => [...prev, obj]);
+          const { data } = await axios.post(
+            "https://631076b736e6a2a04eeef849.mockapi.io/cartItems",
+            obj
+          );
+          setCartItems((prev:any) =>
+            prev.map((item:{parentId:string|null}) => {
+              if (item.parentId  === data.parentId) {
+                return {
+                  ...item,
+                  id: data.id,
+                };
+              }
+              return item;
+            })
+          );
+      }
       }
     } catch (error) {
       console.log("Error with adding card to cart, ", error);
     }
   };
 
-  const isItemAdded = (id) => {
-    return cartItems.some((obj) => Number(obj.parentId) === Number(id));
+  const isItemAdded = (id:null|string) => {
+    if(cartItems) return cartItems.some((obj) => Number(obj.parentId as number|string) === Number(id))
   };
 
-  const onRemoveCartItems = (id) => {
+  const onRemoveCartItems = (id:number|string) => {
     try {
       axios.delete(
         `https://631076b736e6a2a04eeef849.mockapi.io/cartItems/${id}`
       );
-      setCartItems((prev) => prev.filter((el) => Number(el.id) !== Number(id)));
+      setCartItems((prev:any) => prev.filter((el:{id:string}) => Number(el.id as string|null) !== Number(id)));
     } catch (error) {
-      alert(" Owibka udaleniya ", error);
+      alert(" Owibka udaleniya ");
     }
   };
 
@@ -96,10 +115,11 @@ function App() {
         onAddToCart,
         setCartItems
       }}
-    >
+      >
       <div className="wrapper">
         <Routes>
-          <Route exact path="/react-testshop/" element={<Home />} />
+          <Route path="/react-testshop/" element={<Home />} />
+          <Route path="/react-testshop/description/" element={<ItemDescription />} />
           <Route path="/react-testshop/cart/" element={<Cart />} />
           <Route path="/react-testshop/orders/" element={<Orders />} />
         </Routes>
